+++ src/main/java/com/doublew2w/sbs/mybatis/executor/resultset/DefaultResultSetHandler.java
@@ -27,6 +27,7 @@ import lombok.extern.slf4j.Slf4j;
  */
 @Slf4j
 public class DefaultResultSetHandler implements ResultSetHandler {
+  private static final Object NO_VALUE = new Object();
 
   private final Configuration configuration;
   private final MappedStatement mappedStatement;
@@ -80,17 +81,30 @@ public class DefaultResultSetHandler implements ResultSetHandler {
       List<Object> multipleResults,
       ResultMapping parentMapping)
       throws SQLException {
-    if (resultHandler == null) {
-      // 1. 新创建结果处理器
-      DefaultResultHandler defaultResultHandler = new DefaultResultHandler(objectFactory);
-      // 2. 封装数据
-      handleRowValuesForSimpleResultMap(rsw, resultMap, defaultResultHandler, rowBounds, null);
-      // 3. 保存结果
-      multipleResults.add(defaultResultHandler.getResultList());
+    try {
+      if (resultHandler == null) {
+        // 1. 新创建结果处理器
+        DefaultResultHandler defaultResultHandler = new DefaultResultHandler(objectFactory);
+        // 2. 封装数据
+        handleRowValuesForSimpleResultMap(rsw, resultMap, defaultResultHandler, rowBounds, null);
+        // 3. 保存结果
+        multipleResults.add(defaultResultHandler.getResultList());
+      }
+    } finally {
+      // issue #228 (close resultsets)
+      closeResultSet(rsw.getResultSet());
     }
   }
 
-  private void closeResultSet(ResultSet resultSet) {}
+  private void closeResultSet(ResultSet rs) {
+    try {
+      if (rs != null) {
+        rs.close();
+      }
+    } catch (SQLException e) {
+      // ignore
+    }
+  }
 
   /** 处理行记录 */
   private void handleRowValuesForSimpleResultMap(
@@ -119,7 +133,10 @@ public class DefaultResultSetHandler implements ResultSetHandler {
     Object resultObject = createResultObject(rsw, resultMap, null);
     if (resultObject != null && !typeHandlerRegistry.hasTypeHandler(resultMap.getType())) {
       final MetaObject metaObject = configuration.newMetaObject(resultObject);
+      // 自动映射：把每列的值都赋到对应的字段上
       applyAutomaticMappings(rsw, resultMap, metaObject, null);
+      // Map映射：根据映射类型赋值到字段
+      applyPropertyMappings(rsw, resultMap, metaObject, null);
     }
     return resultObject;
   }
@@ -156,6 +173,30 @@ public class DefaultResultSetHandler implements ResultSetHandler {
             // 通过反射工具类设置属性值
             metaObject.setValue(property, value);
           }
+        }
+      }
+    }
+    return foundValues;
+  }
+
+  private boolean applyPropertyMappings(
+      ResultSetWrapper rsw, ResultMap resultMap, MetaObject metaObject, String columnPrefix)
+      throws SQLException {
+    final List<String> mappedColumnNames = rsw.getMappedColumnNames(resultMap, columnPrefix);
+    boolean foundValues = false;
+    final List<ResultMapping> propertyMappings = resultMap.getPropertyResultMappings();
+    for (ResultMapping propertyMapping : propertyMappings) {
+      final String column = propertyMapping.getColumn();
+      if (column != null && mappedColumnNames.contains(column.toUpperCase(Locale.ENGLISH))) {
+        // 获取值
+        final TypeHandler<?> typeHandler = propertyMapping.getTypeHandler();
+        Object value = typeHandler.getResult(rsw.getResultSet(), column);
+        // 设置值
+        final String property = propertyMapping.getProperty();
+        if (value != NO_VALUE && property != null && value != null) {
+          // 通过反射工具类设置属性值
+          metaObject.setValue(property, value);
+          foundValues = true;
         }
       }
     }