+++ src/main/java/com/doublew2w/sbs/mybatis/mapping/ResultMap.java
@@ -3,6 +3,7 @@ package com.doublew2w.sbs.mybatis.mapping;
 import com.doublew2w.sbs.mybatis.session.Configuration;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Locale;
 import java.util.Set;
 import lombok.Getter;
 
@@ -41,7 +42,18 @@ public class ResultMap {
 
     public ResultMap build() {
       resultMap.mappedColumns = new HashSet<>();
+      // 添加 mappedColumns 字段
+      for (ResultMapping resultMapping : resultMap.resultMappings) {
+        final String column = resultMapping.getColumn();
+        if (column != null) {
+          resultMap.mappedColumns.add(column.toUpperCase(Locale.ENGLISH));
+        }
+      }
       return resultMap;
     }
+  }
+
+  public List<ResultMapping> getPropertyResultMappings() {
+    return resultMappings;
   }
 }