+++ src/main/java/com/doublew2w/sbs/mybatis/mapping/ResultMapping.java
@@ -3,6 +3,9 @@ package com.doublew2w.sbs.mybatis.mapping;
 import com.doublew2w.sbs.mybatis.session.Configuration;
 import com.doublew2w.sbs.mybatis.type.JdbcType;
 import com.doublew2w.sbs.mybatis.type.TypeHandler;
+import com.doublew2w.sbs.mybatis.type.TypeHandlerRegistry;
+import java.util.ArrayList;
+import java.util.List;
 import lombok.Getter;
 
 /**
@@ -33,51 +36,43 @@ public class ResultMapping {
   /** 类型处理器 */
   private TypeHandler<?> typeHandler;
 
+  /** 结果标志列表 */
+  private List<ResultFlag> flags;
+
   ResultMapping() {}
 
   public static class Builder {
     private ResultMapping resultMapping = new ResultMapping();
 
-    public Builder(
-        Configuration configuration, String property, String column, TypeHandler<?> typeHandler) {
-      this(configuration, property);
-      resultMapping.column = column;
-      resultMapping.typeHandler = typeHandler;
-    }
-
     public Builder(Configuration configuration, String property, String column, Class<?> javaType) {
-      this(configuration, property);
-      resultMapping.column = column;
-      resultMapping.javaType = javaType;
-    }
-
-    public Builder(Configuration configuration, String property) {
       resultMapping.configuration = configuration;
       resultMapping.property = property;
-    }
-
-    public Builder javaType(Class<?> javaType) {
-      resultMapping.javaType = javaType;
-      return this;
-    }
-
-    public Builder column(String column) {
       resultMapping.column = column;
-      return this;
-    }
-
-    public Builder jdbcType(JdbcType jdbcType) {
-      resultMapping.jdbcType = jdbcType;
-      return this;
-    }
-
-    public ResultMapping build() {
-      return resultMapping;
+      resultMapping.javaType = javaType;
+      resultMapping.flags = new ArrayList<>();
     }
 
     public Builder typeHandler(TypeHandler<?> typeHandler) {
       resultMapping.typeHandler = typeHandler;
       return this;
+    }
+
+    public Builder flags(List<ResultFlag> flags) {
+      resultMapping.flags = flags;
+      return this;
+    }
+
+    public ResultMapping build() {
+      resolveTypeHandler();
+      return resultMapping;
+    }
+
+    private void resolveTypeHandler() {
+      if (resultMapping.typeHandler == null && resultMapping.javaType != null) {
+        Configuration configuration = resultMapping.configuration;
+        TypeHandlerRegistry typeHandlerRegistry = configuration.getTypeHandlerRegistry();
+        resultMapping.typeHandler = typeHandlerRegistry.getTypeHandler(resultMapping.javaType, null);
+      }
     }
   }
 }