+++ src/main/java/com/doublew2w/sbs/mybatis/type/TypeHandlerRegistry.java
@@ -1,6 +1,8 @@
 package com.doublew2w.sbs.mybatis.type;
 
+import java.lang.reflect.Constructor;
 import java.lang.reflect.Type;
+import java.util.Date;
 import java.util.EnumMap;
 import java.util.HashMap;
 import java.util.Map;
@@ -23,6 +25,7 @@ public final class TypeHandlerRegistry {
     register(String.class, new StringTypeHandler());
     register(String.class, JdbcType.CHAR, new StringTypeHandler());
     register(String.class, JdbcType.VARCHAR, new StringTypeHandler());
+    register(Date.class, new DateTypeHandler());
   }
 
   private void register(Type javaType, TypeHandler<?> handler) {
@@ -63,5 +66,36 @@ public final class TypeHandlerRegistry {
     }
     // type drives generics here
     return (TypeHandler<T>) handler;
+  }
+
+  public TypeHandler<?> getMappingTypeHandler(Class<? extends TypeHandler<?>> typeHandlerType) {
+    return ALL_TYPE_HANDLERS_MAP.get(typeHandlerType);
+  }
+
+  /**
+   * 创建一个类型处理器
+   *
+   * @param javaTypeClass java类型
+   * @param typeHandlerClass 处理器类型
+   */
+  @SuppressWarnings("unchecked")
+  public <T> TypeHandler<T> getInstance(Class<?> javaTypeClass, Class<?> typeHandlerClass) {
+    if (javaTypeClass != null) {
+      try {
+        Constructor<?> c = typeHandlerClass.getConstructor(Class.class);
+        return (TypeHandler<T>) c.newInstance(javaTypeClass);
+      } catch (NoSuchMethodException ignored) {
+        // ignored
+      } catch (Exception e) {
+        throw new RuntimeException(
+            "Failed invoking constructor for handler " + typeHandlerClass, e);
+      }
+    }
+    try {
+      Constructor<?> c = typeHandlerClass.getConstructor();
+      return (TypeHandler<T>) c.newInstance();
+    } catch (Exception e) {
+      throw new RuntimeException("Unable to find a usable constructor for " + typeHandlerClass, e);
+    }
   }
 }