+++ src/main/resources/mybatis.sql
@@ -1,14 +1,40 @@
-DROP TABLE user;
-CREATE TABLE user
-(
-    id         bigint    NOT NULL AUTO_INCREMENT COMMENT '自增ID',
-    userId     varchar(9) COMMENT '用户ID',
-    userHead   varchar(16) COMMENT '用户头像',
-    createTime timestamp NULL COMMENT '创建时间',
-    updateTime timestamp NULL COMMENT '更新时间',
-    userName   varchar(64),
-    PRIMARY KEY (id)
-) ENGINE = InnoDB
-  DEFAULT CHARSET = utf8;
-insert into user (id, userId, userHead, createTime, updateTime, userName)
-values (1, '10001', '1_04', '2022-04-13 00:00:00', '2022-04-13 00:00:00', '小傅哥');
+DROP TABLE IF EXISTS `activity`;
+CREATE TABLE `activity` (
+                            `id` bigint(20) NOT NULL AUTO_INCREMENT COMMENT '自增ID',
+                            `activity_id` bigint(20) NOT NULL COMMENT '活动ID',
+                            `activity_name` varchar(64) CHARACTER SET utf8mb4 DEFAULT NULL COMMENT '活动名称',
+                            `activity_desc` varchar(128) CHARACTER SET utf8mb4 DEFAULT NULL COMMENT '活动描述',
+                            `create_time` datetime DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间',
+                            `update_time` datetime DEFAULT CURRENT_TIMESTAMP COMMENT '修改时间',
+                            PRIMARY KEY (`id`),
+                            UNIQUE KEY `unique_activity_id` (`activity_id`)
+) ENGINE=InnoDB AUTO_INCREMENT=4 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin COMMENT='活动配置';
+
+-- ----------------------------
+-- Records of activity
+-- ----------------------------
+BEGIN;
+INSERT INTO `activity` VALUES (1, 100001, '活动名', '测试活动', '2021-08-08 20:14:50', '2021-08-08 20:14:50');
+INSERT INTO `activity` VALUES (3, 100002, '活动名', '测试活动', '2021-10-05 15:49:21', '2021-10-05 15:49:21');
+COMMIT;
+
+-- ----------------------------
+-- Table structure for user
+-- ----------------------------
+DROP TABLE IF EXISTS `user`;
+CREATE TABLE `user` (
+                        `id` bigint(20) NOT NULL AUTO_INCREMENT COMMENT '自增ID',
+                        `userId` varchar(9) DEFAULT NULL COMMENT '用户ID',
+                        `userHead` varchar(16) DEFAULT NULL COMMENT '用户头像',
+                        `createTime` timestamp NULL DEFAULT NULL COMMENT '创建时间',
+                        `updateTime` timestamp NULL DEFAULT NULL COMMENT '更新时间',
+                        `userName` varchar(64) DEFAULT NULL,
+                        PRIMARY KEY (`id`)
+) ENGINE=InnoDB AUTO_INCREMENT=4 DEFAULT CHARSET=utf8;
+
+-- ----------------------------
+-- Records of user
+-- ----------------------------
+BEGIN;
+INSERT INTO `user` VALUES (1, '10001', '1_04', '2022-04-13 00:00:00', '2022-04-13 00:00:00', '叮当猫');
+COMMIT;