+++ src/test/java/com/doublew2w/sbs/mybatis/test/dao/IActivityDao.java
@@ -1,0 +1,12 @@
+package com.doublew2w.sbs.mybatis.test.dao;
+
+import com.doublew2w.sbs.mybatis.test.po.Activity;
+
+/**
+ * @author: DoubleW2w
+ * @date: 2024/9/19 11:33
+ * @project: sbs-mybatis
+ */
+public interface IActivityDao {
+  Activity queryActivityById(Long activityId);
+}