+++ src/test/java/com/doublew2w/sbs/mybatis/test/po/Activity.java
@@ -1,0 +1,33 @@
+package com.doublew2w.sbs.mybatis.test.po;
+
+import java.util.Date;
+import lombok.Data;
+
+/**
+ * @author: DoubleW2w
+ * @date: 2024/9/19 11:33
+ * @project: sbs-mybatis
+ */
+@Data
+public class Activity {
+  /** 自增ID */
+  private Long id;
+
+  /** 活动ID */
+  private Long activityId;
+
+  /** 活动名称 */
+  private String activityName;
+
+  /** 活动描述 */
+  private String activityDesc;
+
+  /** 创建人 */
+  private String creator;
+
+  /** 创建时间 */
+  private Date createTime;
+
+  /** 修改时间 */
+  private Date updateTime;
+}